refactor: Remove deprecated FwbBlockquote component and related files
This commit removes the deprecated FwbBlockquote component and its related files:
- src/components/FwbBlockquote/FwbBlockquote.vue
- src/components/FwbBlockquote/composables/useBlockquoteClasses.ts
- src/components/FwbBlockquote/types.ts

The functionality of the FwbBlockquote component has been replaced by the new FwbBlockquote component located at src/components/Typography/FwbBlockquote.vue.

Note: This commit message follows the established convention of starting with a type (refactor) followed by a concise description of the changes.
