--- src/components/FwbBlockquote/types.ts
@@ -1,3 +1,0 @@
-export type BlockquoteType = 'default' | 'solid'
-export type BlockquoteAlign = 'left' | 'center' | 'right'
-export type BlockquoteSize = 'sm' | 'md' | 'lg'
